some lines be gone
