+++ Makefile
@@ -6,7 +6,7 @@
 #    By: dgarcez- <dgarcez-@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2025/03/11 16:46:46 by dgarcez-          #+#    #+#              #
-#    Updated: 2025/05/16 16:22:26 by dgarcez-         ###   ########.fr        #
+#    Updated: 2025/05/16 19:36:46 by dgarcez-         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -19,12 +19,13 @@ CFLAGS = -Wall -Wextra -Werror -g
 
 SOURCES = minishell.c \
 		  parser/parsing.c \
+          parser/tokenization_utils.c \
+          parser/tokenization.c \
           parser/check_tokens.c \
           parser/expansions.c \
 		  parser/dollar_exp.c \
 		  parser/expansions_utils.c \
-          parser/tokenization_utils.c \
-          parser/tokenization.c \
+		  parser/final_split.c \
 		  parser/create_array.c \
 		  parser/ast_utils.c \
 		  parser/ast_tree.c \
