+++ srcs/parser/expansions.c
@@ -6,7 +6,7 @@
 /*   By: dgarcez- <dgarcez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/11 15:58:38 by dgarcez-          #+#    #+#             */
-/*   Updated: 2025/05/16 13:59:02 by dgarcez-         ###   ########.fr       */
+/*   Updated: 2025/05/16 19:53:37 by dgarcez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -94,161 +94,41 @@ bool	remove_quotes(t_token *token)
 	return (true);
 }
 
-int	unclosed_split_quotes(char *value)
+// int	unclosed_split_quotes(char *value)
+// {
+// 	int		i;
+// 	int		quote;
+// 	char	cquote;
+
+// 	i = 0;
+// 	quote = 0;
+// 	while (value[i])
+// 	{
+// 		if (value[i] == '\"' || value[i] == '\'')
+// 		{
+// 			quote++;
+// 			cquote = value[i];
+// 			i++;
+// 			while (value[i] && value[i] != cquote)
+// 				i++;
+// 			if (value[i] == cquote)
+// 			{
+// 				quote++;
+// 				i++;
+// 			}
+// 		}
+// 		else
+// 			i++;
+// 	}
+// 	return (quote);
+// }
+
+int	new_tokens_amount(t_token *tokens, int i, int j)
 {
-	int		i;
-	int		quote;
-	char	cquote;
-
-	i = 0;
-	quote = 0;
-	while (value[i])
-	{
-		if (value[i] == '\"' || value[i] == '\'')
-		{
-			quote++;
-			cquote = value[i];
-			i++;
-			while (value[i] && value[i] != cquote)
-				i++;
-			if (value[i] == cquote)
-			{
-				quote++;
-				i++;
-			}
-		}
-		else
-			i++;
-	}
-	return (quote);
-}
-
-void	do_count_words_loop_sigma_best_ever_made(int *i, char *s, char quote,
-		int *count)
-{
-	while (s[*i] && (s[*i] == ' ' || (s[*i] >= 9 && s[*i] <= 13)))
-		(*i)++;
-	if (s[*i])
-		(*count)++;
-	while (s[*i] && (s[*i] != ' ' && (s[*i] < 9 || s[*i] > 13)))
-	{
-		if (s[*i] == '\'' || s[*i] == '\"')
-		{
-			quote = s[*i];
-			(*i)++;
-			while (s[*i] && s[*i] != quote)
-				(*i)++;
-			if (s[*i] && s[*i] == quote)
-				(*i)++;
-		}
-		else
-			(*i)++;
-	}
-}
-static int	countwords(char *s)
-{
-	int		i;
-	char	quote;
-	int		count;
-
-	i = 0;
-	count = 0;
-	quote = 0;
-	while (s[i])
-		do_count_words_loop_sigma_best_ever_made(&i, s, quote, &count);
-	return (count);
-}
-
-static void	quote_handle(int *len, char *s, int *i)
-{
-	char	quote;
-
-	quote = s[*i];
-	(*i)++;
-	(*len)++;
-	while (s[*i] && s[*i] != quote)
-	{
-		(*i)++;
-		(*len)++;
-	}
-	if (s[*i] && s[*i] == quote)
-		(*i)++;
-	(*len)++;
-}
-
-static char	*wordalloc(char *s, char c, int *i)
-{
-	char	*word;
-	char	*str;
-	int		len;
-
-	len = 0;
-	while (s[*i] && (s[*i] == c || (s[*i] >= 9 && s[*i] <= 13)))
-		(*i)++;
-	str = s + (*i);
-	while (s[*i] && (s[*i] != c || (s[*i] < 9 && s[*i] > 13)))
-	{
-		if (s[*i] && (s[*i] == '\'' || s[*i] == '\"'))
-			quote_handle(&len, s, i);
-		else
-		{
-			(*i)++;
-			len++;
-		}
-	}
-	word = malloc(sizeof(char) * (len + 1));
-	if (!word)
-		return (NULL);
-	return (ft_strlcpy(word, str, len + 1), word);
-}
-
-char	**ft_arg_split(char *s, char c)
-{
-	char	**result;
-	int		stringnum;
-	int		i;
-	int		index;
-
-	index = 0;
-	if (!s)
-		return (NULL);
-	i = 0;
-	stringnum = countwords(s);
-	result = malloc(sizeof(char *) * (stringnum + 1));
-	if (!result)
-		return (NULL);
-	result[stringnum] = 0;
-	while (i < stringnum)
-	{
-		result[i] = wordalloc(s, c, &index);
-		if (!result[i])
-			return (freetrix(result), NULL);
-		i++;
-	}
-	return (result);
-}
-
-t_token	*expand_strs(t_token *tokens, t_mini *shell)
-{
-	int		i;
-	int		j;
-	int		k;
-	int		amount;
-	t_token	*new_tokens;
 	char	**res;
+	int		amount;
 
-	i = 0;
 	amount = 0;
-	while (tokens[i].type != T_NULL)
-	{
-		if (tokens[i].type != T_PIPE)
-		{
-			if (tokens[i].type != T_HERE_DOC)
-				dollar_expand(&tokens[i], shell);
-		}
-		i++;
-	}
-	i = 0;
 	while (tokens[i].type != T_NULL)
 	{
 		if (!(tokens[i].type >= T_HERE_DOC && tokens[i].type <= T_APPEND_REDIR))
@@ -269,34 +149,42 @@ t_token	*expand_strs(t_token *tokens, t_mini *shell)
 			i++;
 		}
 	}
-	i = 0;
-	k = 0;
+	return (amount);
+}
+
+void	put_new_tokens(t_token *tokens, t_token *new_tokens, int *k, int *i)
+{
+	char	**res;
+	int		j;
+
+	j = -1;
+	res = ft_arg_split(tokens[*i].value, ' ');
+	while (res && res[++j])
+	{
+		new_tokens[*k].value = ft_strdup(res[j]);
+		if (tokens[*i].type == T_PIPE)
+			new_tokens[*k].type = token_type(new_tokens[*k].value, 1);
+		else
+			new_tokens[*k].type = token_type(new_tokens[*k].value, 0);
+		(*k)++;
+	}
+	(*i)++;
+	freetrix(res);
+}
+
+t_token	*create_new_tokens(t_token *tokens, int amount, int i, int k)
+{
+	t_token	*new_tokens;
+
 	new_tokens = ft_calloc(amount + 1, sizeof(t_token));
 	if (new_tokens == NULL)
 		return (NULL);
 	new_tokens[amount].type = T_NULL;
 	new_tokens[amount].value = NULL;
-	printf("amout = %d\n", amount);
 	while (tokens[i].type != T_NULL)
 	{
 		if (!(tokens[i].type >= T_HERE_DOC && tokens[i].type <= T_APPEND_REDIR))
-		{
-			j = -1;
-			res = ft_arg_split(tokens[i].value, ' ');
-			if (res == NULL)
-				amount++;
-			while (res && res[++j])
-			{
-				new_tokens[k].value = ft_strdup(res[j]);
-				if (tokens[i].type == T_PIPE)
-					new_tokens[k].type = token_type(new_tokens[k].value, 1);
-				else
-					new_tokens[k].type = token_type(new_tokens[k].value, 0);
-				k++;
-			}
-			i++;
-			freetrix(res);
-		}
+			put_new_tokens(tokens, new_tokens, &k, &i);
 		else
 		{
 			new_tokens[k].value = ft_strdup(tokens[i].value);
@@ -308,6 +196,29 @@ t_token	*expand_strs(t_token *tokens, t_mini *shell)
 			i++;
 		}
 	}
+	return (new_tokens);
+}
+
+t_token	*expand_strs(t_token *tokens, t_mini *shell)
+{
+	int		i;
+	int		amount;
+	t_token	*new_tokens;
+
+	i = 0;
+	amount = 0;
+	while (tokens[i].type != T_NULL)
+	{
+		if (tokens[i].type != T_PIPE)
+		{
+			if (tokens[i].type != T_HERE_DOC)
+				dollar_expand(&tokens[i], shell);
+		}
+		i++;
+	}
+	i = 0;
+	amount = new_tokens_amount(tokens, 0, 0);
+	new_tokens = create_new_tokens(tokens, amount, 0, 0);
 	i = 0;
 	while (new_tokens[i].type != T_NULL)
 	{
